# swagger/src/foxnox.openapi.example.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Foxnox — password management service",
    "description": "Manages password hashes, verification, rotation, tokens, password policies, and trusted devices for the Foxnox stack.\n\n**Dev credentials.** The `__PWD_<userId>__` placeholders in the `/foxnox/compare` examples are replaced with real plaintext passwords by `scripts/setup-mocks.sh` after it seeds mock accounts via `POST /foxnox/`. If you still see literal `__PWD_*__` tokens below, run `scripts/setup-mocks.sh`."
  },
  "servers": [
    {
      "url": "http://localhost:8100/api",
      "description": "Local dev stack (via Traefik → Gatelin)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "userTrustedDevice": {
        "type": "object",
        "description": "A device the user has marked as trusted (skips 2FA until `expiresAt`). `deviceTokenHash` is INSERT-only and never returned.",
        "properties": {
          "id": { "type": "integer", "format": "int32", "readOnly": true },
          "userId": { "type": "integer", "format": "int32", "minimum": 1 },
          "deviceName": { "type": "string", "minLength": 0, "maxLength": 100, "nullable": true, "example": "MacBook Pro (Chrome)" },
          "ipAddress": { "type": "string", "minLength": 0, "maxLength": 45, "nullable": true, "example": "192.168.1.24" },
          "userAgent": { "type": "string", "nullable": true },
          "expiresAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true },
          "archived": { "type": "boolean", "readOnly": true },
          "archivedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "creatorName": { "type": "string", "readOnly": true, "nullable": true },
          "updatedAt": { "type": "string", "format": "date-time", "nullable": true, "readOnly": true },
          "updaterName": { "type": "string", "readOnly": true, "nullable": true }
        },
        "required": ["userId", "expiresAt"]
      },
...
      "errorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "integer", "example": 400 },
          "message": { "type": "string", "example": "Bad Request" },
          "details": { "type": "string", "example": "Invalid request parameters" }
        }
      },
...
    },
    "responses": {
      "badRequestError": {
        "description": "Missing or invalid parameters in body / path / query",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/errorResponse" } } }
      },
...
      "notFoundError": {
        "description": "No matching row",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/errorResponse" } } }
      },
      "unexpectedError": {
        "description": "Unexpected server error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/errorResponse" } } }
      }
    }
  },
  "paths": {
...
    "/foxnox/devices/": {
      "post": {
        "tags": ["Trusted Device"],
        "summary": "Register trusted devices",
        "description": "The caller must hash the raw device token before submitting; `deviceTokenHash` is INSERT-only, `isPrivate`, and never returned.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rows"],
                "properties": {
                  "rows": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/userTrustedDevice" } }
                }
              },
              "example": {
                "rows": [
                  {
                    "userId": 1,
                    "deviceName": "MacBook Pro (Chrome)",
                    "ipAddress": "192.168.1.24",
                    "userAgent": "Mozilla/5.0 …",
                    "expiresAt": "2026-11-11T22:00:00Z"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered device rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": { "type": "array", "items": { "$ref": "#/components/schemas/userTrustedDevice" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/badRequestError" },
          "5XX": { "$ref": "#/components/responses/unexpectedError" }
        }
      },
      "put": {
        "tags": ["Trusted Device"],
        "summary": "Update trusted devices",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rows"],
                "properties": {
                  "rows": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/userTrustedDevice" } }
                }
              },
              "example": { "rows": [{ "id": 1, "lastUsedAt": "2026-08-11T22:00:00Z" }] }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated device rows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": { "type": "array", "items": { "$ref": "#/components/schemas/userTrustedDevice" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/badRequestError" },
          "404": { "$ref": "#/components/responses/notFoundError" },
          "5XX": { "$ref": "#/components/responses/unexpectedError" }
        }
      }
    },
...
  }
}